requirements.txt
# Core dependencies
streamlit>=1.28.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
supervision>=0.16.0
plotly>=5.17.0

# UI components
streamlit-drawable-canvas>=0.9.3

# Data handling
pandas>=2.0.0
PyYAML>=6.0.0

# Utilities
tqdm>=4.66.0

